a more modern look
a more modern look, updated to api 3.0, full power down

+++ src/SHT30.hpp
@@ -27,7 +27,7 @@ bool GetSHT30Values(MyData &myData)
 {
    M5.SHT30.UpdateData();
    if(M5.SHT30.GetError() == 0) {
-      myData.sht30Temperatur = (int) M5.SHT30.GetTemperature();
+      myData.sht30Temperatur = (float) M5.SHT30.GetTemperature();
       myData.sht30Humidity   = (int) M5.SHT30.GetRelHumidity();
       return true;
    }
